+++ app/src/main/java/com/p2p/wallet/user/model/TokenBid.kt
@@ -1,6 +1,8 @@
 package com.p2p.wallet.user.model
 
+import java.math.BigDecimal
+
 data class TokenBid(
     val symbol: String,
-    val bid: Double
+    val bid: BigDecimal
 )
